Fix the second batch of MSVC warnings
The first cleanup uncovered these because github caps annotations per
job. Placement delete operators for the RandomX vm classes, casts for
the pointer difference and code emitter sizes, plain machine macros
instead of the EVAL_DEFINE trick that upsets the MSVC preprocessor,
a pragma for the flexible array in the SMBIOS struct, an unused catch
variable and one more rapidjson Reserve cast.

+++ src/crypto/randomx/jit_compiler_x86.hpp
@@ -141,7 +141,7 @@ namespace randomx {
 
 		static void emit(const uint8_t* src, size_t count, uint8_t* code, uint32_t& codePos) {
 			memcpy(code + codePos, src, count);
-			codePos += count;
+			codePos += static_cast<uint32_t>(count);
 		}
 
 	public:
